requirements.txt
websockets>=12.0
fastapi>=0.100.0
uvicorn>=0.23.0
httpx>=0.25.0
schedule>=1.2.0
pydantic>=2.0.0
python-dotenv>=1.0.0
